Unifying reconciler
