--- internal/reconciler/reconciler_test.go
@@ -1,184 +1,0 @@
-// Copyright 2020 Coinbase, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package reconciler
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/coinbase/rosetta-sdk-go/types"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestContainsAccountCurrency(t *testing.T) {
-	currency1 := &types.Currency{
-		Symbol:   "Blah",
-		Decimals: 2,
-	}
-	currency2 := &types.Currency{
-		Symbol:   "Blah2",
-		Decimals: 2,
-	}
-	accts := []*AccountCurrency{
-		{
-			Account: &types.AccountIdentifier{
-				Address: "test",
-			},
-			Currency: currency1,
-		},
-		{
-			Account: &types.AccountIdentifier{
-				Address: "cool",
-				SubAccount: &types.SubAccountIdentifier{
-					Address: "test2",
-				},
-			},
-			Currency: currency1,
-		},
-		{
-			Account: &types.AccountIdentifier{
-				Address: "cool",
-				SubAccount: &types.SubAccountIdentifier{
-					Address: "test2",
-					Metadata: map[string]interface{}{
-						"neat": "stuff",
-					},
-				},
-			},
-			Currency: currency1,
-		},
-	}
-
-	t.Run("Non-existent account", func(t *testing.T) {
-		assert.False(t, ContainsAccountCurrency(accts, &AccountCurrency{
-			Account: &types.AccountIdentifier{
-				Address: "blah",
-			},
-			Currency: currency1,
-		}))
-	})
-
-	t.Run("Basic account", func(t *testing.T) {
-		assert.True(t, ContainsAccountCurrency(accts, &AccountCurrency{
-			Account: &types.AccountIdentifier{
-				Address: "test",
-			},
-			Currency: currency1,
-		}))
-	})
-
-	t.Run("Basic account with bad currency", func(t *testing.T) {
-		assert.False(t, ContainsAccountCurrency(accts, &AccountCurrency{
-			Account: &types.AccountIdentifier{
-				Address: "test",
-			},
-			Currency: currency2,
-		}))
-	})
-
-	t.Run("Account with subaccount", func(t *testing.T) {
-		assert.True(t, ContainsAccountCurrency(accts, &AccountCurrency{
-			Account: &types.AccountIdentifier{
-				Address: "cool",
-				SubAccount: &types.SubAccountIdentifier{
-					Address: "test2",
-				},
-			},
-			Currency: currency1,
-		}))
-	})
-
-	t.Run("Account with subaccount and metadata", func(t *testing.T) {
-		assert.True(t, ContainsAccountCurrency(accts, &AccountCurrency{
-			Account: &types.AccountIdentifier{
-				Address: "cool",
-				SubAccount: &types.SubAccountIdentifier{
-					Address: "test2",
-					Metadata: map[string]interface{}{
-						"neat": "stuff",
-					},
-				},
-			},
-			Currency: currency1,
-		}))
-	})
-
-	t.Run("Account with subaccount and unique metadata", func(t *testing.T) {
-		assert.False(t, ContainsAccountCurrency(accts, &AccountCurrency{
-			Account: &types.AccountIdentifier{
-				Address: "cool",
-				SubAccount: &types.SubAccountIdentifier{
-					Address: "test2",
-					Metadata: map[string]interface{}{
-						"neater": "stuff",
-					},
-				},
-			},
-			Currency: currency1,
-		}))
-	})
-}
-
-func TestExtractAmount(t *testing.T) {
-	var (
-		currency1 = &types.Currency{
-			Symbol:   "curr1",
-			Decimals: 4,
-		}
-
-		currency2 = &types.Currency{
-			Symbol:   "curr2",
-			Decimals: 7,
-		}
-
-		amount1 = &types.Amount{
-			Value:    "100",
-			Currency: currency1,
-		}
-
-		amount2 = &types.Amount{
-			Value:    "200",
-			Currency: currency2,
-		}
-
-		balances = []*types.Amount{
-			amount1,
-			amount2,
-		}
-
-		badCurr = &types.Currency{
-			Symbol:   "no curr",
-			Decimals: 100,
-		}
-	)
-
-	t.Run("Non-existent currency", func(t *testing.T) {
-		result, err := ExtractAmount(balances, badCurr)
-		assert.Nil(t, result)
-		assert.EqualError(t, err, fmt.Errorf("could not extract amount for %+v", badCurr).Error())
-	})
-
-	t.Run("Simple account", func(t *testing.T) {
-		result, err := ExtractAmount(balances, currency1)
-		assert.Equal(t, amount1, result)
-		assert.NoError(t, err)
-	})
-
-	t.Run("SubAccount", func(t *testing.T) {
-		result, err := ExtractAmount(balances, currency2)
-		assert.Equal(t, amount2, result)
-		assert.NoError(t, err)
-	})
-}
